Simulating a normal data set \(S = (X,Y)\), where \(Y = X^T \beta\).

Creates a data set \(S = (X,Y)\) where the columns of \(X\) are sampled from an independent Gaussian distribution with mean \(\mu_i\) and standard deviation \(\sigma_i\), i.e. \(N(\mu_i, \sigma_i^2)\). The response \(Y\) is given by \(Y = X\beta\). The final dimension will be \(n \times (p + 1)\), with the number of data points \(n\) to be specified.

generate_XY(
  n = 100,
  mu = rep(0, 10),
  sigma = rep(1, 10),
  beta_coefficients = 1:10
)

Arguments

n
desired number of data points in the data set.

mu
a \(p\)-dimensional vector of means for \(\mu\).

sigma
a \(p\)-dimensional vector of non-negative standard deviations for \(\sigma\).

beta_coefficients
a \(p\)-dimensional vector of coefficients for \(\beta\).

Value

An \(n \times (p+1)\) dimensional data frame given by \(S = (X,Y)\). In the base case, the columns \(X_i\) are sampled from \(N(0,1)\) and the coefficients are all 1. We also have \(n = 100\) and \(p = 10\), with beta-coefficients 1 to 10.

Examples

generate_XY()
#> # A tibble: 100 × 11
#>         X1      X2      X3      X4     X5      X6     X7      X8      X9    X10
#>      <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>
#>  1 -0.262   0.0535 -1.19   -0.290   0.261  0.415   0.144 -0.954  -0.285  -0.155
#>  2 -1.42   -0.125   2.25   -1.25   -2.42  -0.0481 -0.718 -0.0800 -0.648  -1.09 
#>  3 -0.514   0.248   1.55   -0.925   1.15   0.198   0.206  1.06   -1.27    0.550
#>  4  0.772  -0.627   0.323  -0.352   1.21  -0.389  -0.367  1.35   -0.175   1.39 
#>  5  1.40    0.683   0.0228 -0.0416 -0.432 -0.452   0.788  0.233  -2.20   -1.59 
#>  6 -0.0158  0.590  -0.249   0.729   0.107 -0.139  -0.311  0.334   0.580   2.07 
#>  7 -1.02   -0.815  -2.17    0.293   0.486 -1.33   -0.859 -0.765   0.0628 -1.70 
#>  8 -2.12   -0.346  -1.34    3.59   -1.55   0.754   1.09   0.769   1.78    0.861
#>  9  0.150   0.0565 -0.980   0.756  -1.08   1.36   -0.785 -0.215   1.38   -0.380
#> 10 -0.521  -0.567  -1.79    0.400  -0.516 -1.22   -1.22   0.383   1.63    1.85 
#> # … with 90 more rows, and 1 more variable: Y <dbl>
generate_XY(n = 40, mu = 1:10, sigma = rep(1, 10), beta_coefficients = 1:10)
#> # A tibble: 40 × 11
#>         X1    X2    X3    X4    X5    X6    X7    X8    X9   X10     Y
#>      <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  0.322  2.12   3.46  3.56  5.34  6.56  7.09  8.52  9.14 10.0   396.
#>  2  1.84   2.10   2.27  1.82  4.68  5.82  6.47  7.41  7.96  8.88  343.
#>  3  1.04   3.36   4.32  3.32  4.05  6.00  6.06  6.82 10.4   9.67  377.
#>  4  1.98   1.16   3.46  4.73  5.12  6.59  5.84  9.20  9.21  8.72  383.
#>  5  2.50   0.935  2.26  2.76  7.21  6.18  8.46  7.09  8.58 10.5   394.
#>  6  2.38   4.11   5.14  3.41  6.30  7.40  5.77  6.45  8.49 11.3   397.
#>  7  1.18   2.51   2.66  5.11  4.47  6.83  8.29  8.41  8.52 10.6   406.
#>  8  0.395  3.20   2.93  3.50  6.15  6.55  7.52  8.87  8.98  9.99  404.
#>  9  0.994  3.05   2.91  4.35  5.83  6.55  7.04  6.14  8.76 10.7   386.
#> 10 -0.0309 2.72   3.36  5.75  6.16  5.46  7.01  8.21  8.96 10.4   401.
#> # … with 30 more rows