Simulating a causal data set \(S = (X,Y_i, T, Y_{obs})\) with multiple potential outcomes.

Creates a causal data set \(S = (X, Y_i, T, Y_{obs})\) for causal inference. The \(p\) columns of \(X\) are sampled from an independent Gaussian distribution with mean \(\mu_i\) with standard deviation \(\sigma_i\), i.e. \(N(\mu_i, \sigma_i^2)\). A treatment \(T\) is sampled, where more than 2 treatments are possible. The observations \(Y_i\) correspond to the outcome if the treatment \(i\) is applied. The outcome \(Y = X^T \beta\) is assumed to depend on \(X\) in a linear fashion, and the treatment effect of treatment \(i\) is additive. See Causality (Pearl 2009) for further details and a general introduction to causal inference.

causal_XTY_multiple(
  n = 100,
  mu = rep(0, 4),
  sigma = rep(1, 4),
  beta_coefficients = 1:4,
  treatment_prob = rep(0.25, 4),
  treatment_effect = c(10, 20, 30, 40)
)

Arguments

n
desired number of data points in the data set.

mu
a \(p\)-dimensional vector of means for \(\mu\).

sigma
a \(p\)-dimensional vector of non-negative standard deviations for \(\sigma\).

beta_coefficients
a \(p\)-dimensional vector of coefficients for \(\beta\).

treatment_prob
a probability vector with weights summing to 1, corresponding to the probability of treatment.

treatment_effect
a vector corresponding to the additive treatment effect of each treatment on the outcome \(Y\).

Value

A causal data set \(S = (X,Y_i, T, Y_{obs})\) with multiple potential outcomes. In the base case, the columns \(X_i\) are sampled from \(N(0,1)\) and the coefficients are all 1. We also have \(n = 100\), \(p = 4\), with beta-coefficients 1 to 10, where \(p\) corresponds to the number of columns in \(X\). The treatment probabilities are equally likely.

Examples

causal_XTY_multiple()
#> # A tibble: 100 × 11
#>        X1     X2     X3      X4     Y    Y1    Y2    Y3    Y4 treatment
#>     <dbl>  <dbl>  <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>     <int>
#>  1  0.718 -1.05   1.76   0.0289  4.01 14.0   24.0  34.0  44.0         3
#>  2  0.539  0.661  2.03  -0.396   6.36 16.4   26.4  36.4  46.4         2
#>  3 -1.06  -0.823  0.428 -0.695  -4.20  5.80  15.8  25.8  35.8         3
#>  4 -0.541  0.437 -0.566 -1.49   -7.33  2.67  12.7  22.7  32.7         2
#>  5 -0.715  0.372 -0.834  1.44    3.31 13.3   23.3  33.3  43.3         4
#>  6 -0.179 -1.65   0.398 -0.347  -3.67  6.33  16.3  26.3  36.3         3
#>  7  0.165 -1.92  -1.47  -0.0403 -8.27  1.73  11.7  21.7  31.7         1
#>  8 -0.727  0.381 -0.395  1.25    3.83 13.8   23.8  33.8  43.8         2
#>  9 -0.538  1.38  -0.131 -1.35   -3.56  6.44  16.4  26.4  36.4         2
#> 10 -0.600  0.826 -2.00  -0.574  -7.24  2.76  12.8  22.8  32.8         3
#> # … with 90 more rows, and 1 more variable: Y_observed <dbl>
causal_XTY_multiple(n = 40, mu = rep(2, 7), sigma = 1:7,
                    beta_coefficients = 1:7,
                    treatment_prob = c(0.4, 0.1, 0.1, 0.2, 0.2),
                    treatment_effect = 1:5)
#> # A tibble: 40 × 15
#>       X1     X2     X3     X4     X5     X6     X7      Y     Y1     Y2     Y3
#>    <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 1.93   4.20  -1.41   6.78   0.431  3.82   5.47   96.6   97.6   98.6   99.6 
#>  2 2.95   3.34   1.16   5.65   0.568 -0.859 11.1   111.   112.   113.   114.  
#>  3 1.62   2.06  -0.682 -1.15  -2.02   0.552 -9.31  -72.9  -71.9  -70.9  -69.9 
#>  4 1.58   6.74   2.41   0.356  1.53   6.94  -6.75   25.7   26.7   27.7   28.7 
#>  5 1.53   7.17  -0.247  3.91   1.63  -7.33  13.1    86.4   87.4   88.4   89.4 
#>  6 0.310  4.42   3.55   2.74   5.57   2.56  10.8   149.   150.   151.   152.  
#>  7 1.05  -0.502  1.42  -3.30   2.62  -0.202  0.534   6.77   7.77   8.77   9.77
#>  8 1.66   2.80   2.09   6.69   8.48   1.22  -2.47   72.7   73.7   74.7   75.7 
#>  9 2.36   3.87   3.08   2.93  -3.57   4.45   4.79   73.4   74.4   75.4   76.4 
#> 10 2.05   1.05   1.91   3.93  -2.21  -1.50  11.4    85.3   86.3   87.3   88.3 
#> # … with 30 more rows, and 4 more variables: Y4 <dbl>, Y5 <dbl>,
#> #   treatment <int>, Y_observed <dbl>