Simulating a normal data set \(S = (X)\).

Creates a data set \(S = (X)\) where the columns of \(X\) are sampled from an independent Gaussian distribution with mean \(\mu_i\) and standard deviation \(\sigma_i\), i.e. \(N(\mu_i, \sigma_i^2)\). The final dimension will be \(n \times p\), with the number of data points \(n\) to be specified.

generate_X(n = 100, mu = rep(0, 10), sigma = rep(1, 10))

Arguments

n
desired number of data points in the data set.

mu
a \(p\)-dimensional vector of means for \(\mu\).

sigma
a \(p\)-dimensional vector of non-negative standard deviations for \(\sigma\).

Value

An \(n \times p\) dimensional data frame given by \(S = (X)\). In the base case, the columns \(X_i\) are sampled from \(N(0,1)\), \(n = 100\) and \(p = 10\).

Examples

generate_X()
#> # A tibble: 100 × 10
#>         X1      X2      X3      X4      X5      X6     X7     X8     X9     X10
#>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1  1.21   -0.974  -1.68   -1.56    0.237   0.413   0.853 -0.816  0.703 -0.349 
#>  2  0.0320 -1.25    0.136   0.698  -2.32   -0.0992  1.89   0.216 -1.54  -0.545 
#>  3  0.180  -0.0463  0.498   0.0382  0.0790 -0.196  -0.867 -1.14  -1.14  -0.0205
#>  4  0.0772  1.14    1.33    0.235  -0.429   0.621   0.714  0.140  0.371 -0.339 
#>  5  0.558   1.26   -0.0230  1.02   -0.373  -0.0475  0.614 -0.662  0.249  0.186 
#>  6 -0.167  -0.549  -0.192   1.48   -0.0449 -1.10    0.305 -0.534  1.24   0.346 
#>  7 -0.228   0.682  -0.441  -0.0484  1.40    0.598  -0.926 -1.41  -0.372 -0.793 
#>  8  0.243  -0.930  -1.94   -0.595  -0.301  -0.810   1.36   1.22   0.909 -0.749 
#>  9 -0.928  -0.975   1.42   -0.474  -0.966   0.819   0.204  0.426  0.772 -1.85  
#> 10  1.52   -0.0271  0.596  -0.326   1.68    1.10    1.29  -0.121  0.367 -1.56  
#> # … with 90 more rows
generate_X(n = 40, mu = 1:10, sigma = rep(1, 10))
#> # A tibble: 40 × 10
#>       X1    X2    X3    X4    X5    X6    X7    X8    X9   X10
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 1.18  1.45   4.07  4.76  5.55  6.81  6.68  8.61  8.94  9.30
#>  2 0.910 4.12   3.66  3.24  3.65  6.03  8.17  8.80  7.36 11.1 
#>  3 0.887 2.24   2.88  5.06  4.36  6.66  6.10  8.32 10.2   8.03
#>  4 1.16  1.48   2.07  4.57  6.05  5.78  7.99  8.10  8.59  8.76
#>  5 1.56  3.16   4.52  5.45  4.68  5.98  6.87  7.34  8.91 11.0 
#>  6 1.85  1.37   2.92  5.93  5.03  8.09  5.56  8.03  9.47 10.6 
#>  7 0.805 2.86   2.63  4.38  6.06  7.50  7.03  7.35  8.17  8.96
#>  8 0.957 0.318  3.61  4.46  3.41  5.69  4.88  8.91  9.40 11.6 
#>  9 0.471 0.363  4.53  3.69  3.42  7.58  8.67  7.95  9.24 10.1 
#> 10 0.401 3.43   2.92  3.80  5.11  4.85  7.58  6.82  9.63 12.3 
#> # … with 30 more rows