Simulating a normal data set \(S\) that includes categorical variables.

Creates a data set \(S = (X_1, X_2)\) where the columns of \(X_1\) are sampled from an independent Gaussian distribution with mean \(\mu_i\) and standard deviation \(\sigma_i\), i.e. \(N(\mu_i, \sigma_i^2)\), and the columns of \(X_2\) are categorical, sampled with replacement from a given number of categories (indexed by numbers). The final dimension will be \(n \times (p_1 + p_2)\), where \(p_1\) is the number of columns in \(X_1\) and \(p_2\) is the number of columns in \(X_2\), with the number of data points \(n\) to be specified.

generate_X_cat(
  n = 100,
  mu = rep(0, 10),
  sigma = rep(1, 10),
  no_of_cat = c(4, 5)
)

Arguments

n
desired number of data points in the data set.

mu
a \(p_1\)-dimensional vector of means for \(\mu\).

sigma
a \(p_1\)-dimensional vector of non-negative standard deviations for \(\sigma\).

no_of_cat
a \(p_2\)-dimensional vector where the entries indicate the number of categories desired for each column of \(X_2\).

Value

An \(n \times p\) dimensional data frame given by \(S = (X_1, X_2)\). In the base case, the columns of \(X_1\) are sampled from \(N(0,1)\), \(n = 100\) and \(p = 10\), and two additional categorical columns of \(X_2\) are added.

Examples

generate_X_cat()
#> # A tibble: 100 × 12
#>         X1      X2      X3     X4      X5      X6     X7     X8      X9    X10
#>      <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1  0.782   0.825  -0.472  -0.127  1.09   -2.69   -0.570 -1.04   0.308   0.776
#>  2 -0.957  -0.0404 -0.387  -0.113  1.16    0.0348  0.654  0.951  0.713  -2.04 
#>  3  0.0914 -0.605  -0.0329  1.81  -1.51    0.363  -0.512  0.929  1.02   -0.596
#>  4 -0.416  -0.479  -1.59   -0.916  0.114  -0.332  -0.133  1.11  -2.32   -0.295
#>  5 -0.661  -1.63   -1.63    0.326 -0.162   0.344   0.229  0.487  0.898  -1.01 
#>  6  0.684   1.10   -0.619   1.73   0.552   0.274  -0.427  0.324 -0.290  -1.07 
#>  7  0.892  -1.13   -0.429   0.619  0.349   0.294  -1.15  -0.336 -0.751  -2.95 
#>  8 -0.727   0.749   2.14   -0.244 -0.0627 -1.00   -0.626  0.393  0.741  -2.80 
#>  9  1.02    0.507   1.38    0.304 -0.483  -1.12    0.800  0.735  0.0934 -0.113
#> 10 -0.396   0.149  -0.982  -0.156  2.00   -1.54   -0.974  0.932 -1.82   -0.244
#> # … with 90 more rows, and 2 more variables: X11 <fct>, X12 <fct>
generate_X_cat(n = 40, mu = 1:10, sigma = rep(1, 10), no_of_cat = c(2,3))
#> # A tibble: 40 × 12
#>         X1    X2    X3    X4    X5    X6    X7    X8    X9   X10 X11   X12  
#>      <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <fct> <fct>
#>  1 -1.21   1.54   3.17  4.20  6.57  4.58  7.20  7.12  6.84  9.97 A     B    
#>  2 -0.284  1.85   2.87  5.23  4.12  5.01  5.91  8.48 10.2  10.8  B     B    
#>  3  2.02   0.975  2.80  3.59  5.32  6.65  7.84  6.27  8.04  8.14 A     C    
#>  4  1.00   1.77   3.95  3.83  5.03  6.25  5.42  8.35  8.40 10.7  A     B    
#>  5  1.98   2.21   2.21  4.43  3.90  4.62  8.82  8.48  6.98  7.66 A     C    
#>  6 -0.0637 1.84   4.13  3.36  4.07  6.65  6.96  9.10  8.50 10.8  B     B    
#>  7  0.884  1.91   3.33  5.49  6.03  4.29  6.08  7.89 11.3  10.9  A     B    
#>  8  2.74   4.04   2.53  3.48  5.51  4.89  7.65  7.39  8.73 10.7  A     A    
#>  9  0.0442 1.96   3.74  5.29  6.55  4.51  5.94  7.02  8.31  9.50 A     A    
#> 10  1.52   1.35   4.11  2.12  6.80  5.13  6.64  8.31  9.65  8.52 B     C    
#> # … with 30 more rows